+++ src/rtc/rtc_connection.h
@@ -47,6 +47,7 @@ class RTCConnection {
 
   void SetEncodingParameters(
       std::vector<webrtc::RtpEncodingParameters> encodings);
+  void ResetEncodingParameters();
 
   rtc::scoped_refptr<webrtc::PeerConnectionInterface> GetConnection() const;
 
@@ -63,6 +64,8 @@ class RTCConnection {
   RTCMessageSender* sender_;
   std::unique_ptr<PeerConnectionObserver> observer_;
   rtc::scoped_refptr<webrtc::PeerConnectionInterface> connection_;
+  std::vector<webrtc::RtpEncodingParameters> encodings_;
+  std::string mid_;
 };
 
 #endif